// backend/src/utils/enums.ts
export enum call_status {
    waiting = 'waiting',
    active = 'active',
    on_hold = 'on_hold',
    ended = 'ended'
}
export enum event_names {
    call_initiated = 'call_initiated',
    call_routed = 'call_routed',
    call_answered = 'call_answered',
    call_hold = 'call_hold',
    call_ended = 'call_ended'
}